Change module to contain info on available designs in module
This is similarly to DesignExtensions in eZ Publish, but with the additional value that:
1. You can choose not to use a design from an module even if you enable it (but this would be confusing, and error prone)
2. But more importantly the code can avoid more file stat calls to second guess the locations of designs

Considering adding more meta data like this on Module class as well, like:
* translations
* license
* dependencies
* version info
* copyright

From there it might be worthwhile to figure out how to connect the independent features of the module (field type, controller) to the Module class for gui use.
But that can be accomplished by parsing service.ini provided in module as well.